--- BemPHP/VariablesList.php
@@ -1,9 +1,0 @@
-<?php
-use BemPHP\BlocksStorage;
-use BemPHP\Tree;
-
-   $b_logger = Tree::html('b-logger');
-   $b_logger__button = Tree::html('b-logger__button');
-   $b_logger__text_box = Tree::html('b-logger__text-box');
-
-?>
