--- BemPHP/blocks/b-bemphp-logo/b-bemphp-logo.css
@@ -1,10 +1,0 @@
-.b-bemphp-logo{
-    position: absolute;
-    height: 332px;
-    width: 332px;
-    background-image: url('/BemPHP/blocks/b-bemphp-logo/b-bemphp-logo.png');
-    top: 50%;
-    left: 50%;
-    margin-left: -166px;
-    margin-top: -166px;
-}
