--- BemPHP/blocks/b-logger/b-logger.css
@@ -1,21 +1,0 @@
-.b-logger{
-    border-radius: 5px;
-    position: fixed;
-
-
-
-    right: 10px;
-    bottom: 10px; 
-    padding: 5px;
-    background: #ffedcd; 
-    box-shadow: 0.2em 0.2em 5px rgba(111,111,111,0.3);
-    opacity: 0.1;
-
-    transition:all 0.2s ease-in-out;
-    -webkit-transition:all 0.2s ease-in-out; /** Chrome & Safari **/
-    -moz-transition:all 0.2s ease-in-out; /** Firefox **/
-    -o-transition:all 0.2s ease-in-out;
-}
-.b-logger:hover{
-    opacity: 1;
-}
