--- BemPHP/blocks/b-logger/b-logger__button/b-logger__button.css
@@ -1,32 +1,0 @@
-.b-logger__button {
-
-    /*position: absolute;*/
-
-    background-color:#eedcbc;
-    border-radius: 7px;
-    margin-left: 0px;
-    margin-right: 0px;
-
-
-    height:17px;
-    width: 300px;
-
-    cursor:pointer;
-
-    font-size: 16px;
-    text-align: center;
-    vertical-align: middle;
-    color: #e5d3b4;
-    vertical-align: middle;
-    padding:0 0px;
-    /*line-height: 20px;*/
-
-
-
-
-}
-
-.b-logger__button:hover {
-    background-color: #e8d6b7;
-    color: #d0a25b;
-}
