--- BemPHP/blocks/b-logger/b-logger__button/b-logger__button__arrow_down/b-logger__button__arrow_down.css
@@ -1,3 +1,0 @@
-.b-logger__button__arrow_down:after {
-    content: '\25bc';
-}
