--- BemPHP/blocks/b-logger/b-logger__button/b-logger__button__arrow_up/b-logger__button__arrow_up.css
@@ -1,5 +1,0 @@
-.b-logger__button__arrow_up:after {
-    position: relative;
-    bottom: 3px;
-    content: '\25b2';
-}
