--- BemPHP/blocks/b-logger/b-logger__msg/b-logger__msg.css
@@ -1,6 +1,0 @@
-.b-logger__msg
-{
-    font-family: sans-serif;
-    font-size: 11px;
-    color: #000000;
-}
