--- BemPHP/blocks/b-logger/b-logger__msg/b-logger__msg.php
@@ -1,9 +1,0 @@
-<?php
-use BemPHP\BlocksStorage;
-
-BlocksStorage::createBlock('b-logger__msg')
-                ->setBlockDir(__DIR__)
-                ->loadCSSFile()
-                ->setTag('p')
-;
-?>
