--- BemPHP/blocks/b-logger/b-logger__msg/b-logger__msg_error/b-logger__msg_error.css
@@ -1,4 +1,0 @@
-.b-logger__msg_error {
-    color:#CE0000 !important;
-    font-weight: bold;
-}
