--- BemPHP/blocks/b-logger/b-logger__msg/b-logger__msg_info/b-logger__msg_info.css
@@ -1,3 +1,0 @@
-.b-logger__msg_info{
-    color:#00008B !important;
-}
