--- BemPHP/blocks/b-logger/b-logger__msg/b-logger__msg_warn/b-logger__msg_warn.css
@@ -1,4 +1,0 @@
-.b-logger__msg_warn{
-    color:#a53f00 !important;
-    font-weight: bold;
-}
