--- BemPHP/blocks/b-logger/b-logger__text-box/b-logger__text-box.css
@@ -1,12 +1,0 @@
-.b-logger__text-box {
-    /*position: absolute;*/
-
-    width: 300px;
-    height: 300px;
-
-    top : 25px;
-    left: 5px;
-    right: 2px;
-    bottom: 5px;
-    overflow-y: auto;
-}
