+++ src/Notificator.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace BemPHP;
+
+/** интерфейс для оповещения сервисов логирования
+ * Class Notifycator
+ * @package BemPHP
+ */
+interface Notifycator{
+    function notify($logStorage);
+};
