+++ src/VariablesList.php
@@ -1,0 +1,8 @@
+<?php
+
+use BemPHP\BlocksStorage;
+use BemPHP\Tree;
+
+$b_logger = Tree::html('b-logger');
+$b_logger__button = Tree::html('b-logger__button');
+$b_logger__text_box = Tree::html('b-logger__text-box');
